Moodle 2.4 with first version of a modified mod_data

+++ html/web/admin/timezone.php
@@ -1,42 +1,55 @@
-<?php   // $Id: timezone.php,v 1.8.2.1 2008/05/02 04:07:28 dongsheng Exp $
+<?php
 
     require_once('../config.php');
 
-    $zone = optional_param('zone', '', PARAM_PATH); //not a path, but it looks like it anyway
+    $zone = optional_param('zone', '', PARAM_RAW);
+
+    if (!is_numeric($zone)) {
+         //not a path, but it looks like it anyway
+         $zone = clean_param($zone, PARAM_PATH);
+    }
+
+    $PAGE->set_url('/admin/timezone.php');
+    $PAGE->set_context(context_system::instance());
 
     require_login();
 
-    require_capability('moodle/site:config', get_context_instance(CONTEXT_SYSTEM));
+    require_capability('moodle/site:config', context_system::instance());
 
     $strtimezone = get_string("timezone");
     $strsavechanges = get_string("savechanges");
     $strusers = get_string("users");
     $strall = get_string("all");
 
-    print_header($strtimezone, $strtimezone, build_navigation(array(array('name' => $strtimezone, 'link' => null, 'type' => 'misc'))));
+    $PAGE->set_title($strtimezone);
+    $PAGE->set_heading($strtimezone);
+    $PAGE->navbar->add($strtimezone);
+    echo $OUTPUT->header();
 
-    print_heading("");
+    echo $OUTPUT->heading("");
 
-    if (!empty($zone) and confirm_sesskey()) {
-        $db->debug = true;
+    if (data_submitted() and !empty($zone) and confirm_sesskey()) {
         echo "<center>";
-        execute_sql("UPDATE {$CFG->prefix}user SET timezone = '$zone'");
-        $db->debug = false;
+        $DB->execute("UPDATE {user} SET timezone = ?", array($zone));
         echo "</center>";
 
         $USER->timezone = $zone;
+        $current = $zone;
+        echo $OUTPUT->notification('Timezone of all users changed', 'notifysuccess');
+    } else {
+        $current = 99;
     }
 
     require_once($CFG->dirroot.'/calendar/lib.php');
     $timezones = get_list_of_timezones();
 
-    echo '<center><form action="timezone.php" method="get">';
-    echo "$strusers ($strall): ";
-    choose_from_menu ($timezones, "zone", 99, get_string("serverlocaltime"), "", "99");
-    echo "<input type=\"hidden\" name=\"sesskey\" value=\"$USER->sesskey\" />";
-    echo "<input type=\"submit\" value=\"$strsavechanges\" />";
+    echo '<center><form action="timezone.php" method="post">';
+    echo html_writer::label($strusers . ' (' . $strall . '): ', 'menuzone');
+    echo html_writer::select($timezones, "zone", $current, array('99'=>get_string("serverlocaltime")));
+    echo "<input type=\"hidden\" name=\"sesskey\" value=\"".sesskey()."\" />";
+    echo '<input type="submit" value="'.s($strsavechanges).'" />';
     echo "</form></center>";
 
-    print_footer();
+    echo $OUTPUT->footer();
 
-?>
+
